// src/domain/interfaces/lists/list.ts
'use strict';

import { EUserRole } from '../../enum/role.enum';
import { Registable } from '../auditable';

export interface ListAdd<T> {
    readonly items?: T[];
    readonly itemsId: string[];
    readonly title: string;
    readonly readRole: EUserRole;
}

export interface List<T> extends ListAdd<T>, Registable {
    readonly items: T[];
}